Update patch set 2
Patch Set 2:

(1 comment)

> Patch Set 2:
> 
> (2 comments)
> 
> questions only

> 1. this is old code, but why is this unlikely?
Yes，the code of adjust_lru is indeed duplicate of the _mdcache_lru_ref, i will merge the duplicate code. Thank you very much.

> 2. there is only 1 root object, why does this code make it the common case (or am I misreading)?
In addition to the root object '/', we also need to consider the export root object and NFSv4 pseduo root object. There is not only 1 here, it should be related to the number of configured export path.

Patch-set: 2