Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)
> 
> comment updated

the problem i described in the commit msg is exactly what i found from the nfs client and ganesha log. the fundamental problem is the resend request found that the previous request with the same xid is still being processed and drop the request, while the original request with the old connection context is not able to reply to the nfs client, so the nfs client hang permanently.

with respect to your concern about reply twice problems, the alternative approach i am able to take is only retry and reply the resend request when the original one reply failed due to XPRT_DESTROYED. 
how's that sounds?

Patch-set: 1