Update patch set 3
Patch Set 3:

(4 comments)

I think this is an important concept, but I'm not sure this is the best way. Perhaps a better way would be a counter on an inode of the number of exports for which it is the root of an export. I think the counter would have to be atomic unless it is only changed while holding the content lock.

Patch-set: 3
CC: Gerrit User 1003571 <1003571@d5d70762-12d0-45a1-890d-524b12d3f735>