Update patch set 3
Patch Set 3:

But I still think it might be better to do something to mark the mdcache nodes that are export roots, and since there can be more than one export with the same root, the best thing is for that mark to be a counter.

I wonder if we actually should take export root nodes OUT of the LRU completely so long as the count is non-zero? There's really no need to promote and demote them, it's just extra activity.

Note that in the old cache_inode design, they were indeed sort of kept separate because they were pinned (and maybe that DID actually keep them out of the LRU, I don't recall how pinning worked).

Patch-set: 3