Update patch set 2
Patch Set 2:

When we implemented this, it was decided to truncate the name as there could be very large no. of users so DBUS can run into issues like overflow on message limits etc. The value of user->uname.len is already the MIN out of the two, as 1024 is the maximum.

Patch-set: 2