MDCACHE: root_obj must be released before last unexport
root object must be released before last export is released.
If last export of root_obj is released, just left root object
handle in cache, once the export resource is used through
root_obj->export, NULL pointer used cause to crash such as
obj->export->cmount in deconstruct_handle() of FSAL_CEPH.
So can not remove cih_remove_checked() in mdcache_unexport.

Another case, if the object entry to be root of multiple exports,
here should not unhash and set entry as invalid when the first
export unmount, but unhash in last unexport. So we need to judge.

Change-Id: I0068441f94b6185f3b072391edb8c5ec29ea60ca
Signed-off-by: sepia-liu <liuwei_coder@163.com>