Always set the wcc attributes in NFSv3 setattr replies
A special case was added in 4a3af6df1b46075b80f33e43e3f19fd768e6c43a,
as the wcc was set using attribute information that could be
incomplete if only the file size was modified. In this case, the reply
would be marked as containing no wcc data at all.

The attribute information for the wcc is now fetched using getattr
calls to the FSAL layer and the check should no longer be needed.

Signed-off-by: David Hedberg <david.hedberg@zavann.se>
Change-Id: If8584d1b0ddb27ca200af9e85cf2ff11c2d28c2a