[GPFS] Fix to let GPFS upcall thread exit at ganesha startup
Currently during startup if GPFSFSAL_UP_Thread() gets called, it
waits in nfs_init_wait() for init to complete. But if during startup
gpfs_unclaim_filesystem() gets called, then the main thread may hang
at pthread_join() for GPFS upcall thread, if the thread is waiting in
nfs_init_wait().
To avoid this made use of 'stop_thread' variable in
'struct gpfs_filesystem', which is set to 'true' in
gpfs_unclaim_filesystem(). And after that removed call to pthread_join()
to avoid waiting for GPFS upcall thread to exit. The 'stop_thread'
variable value is checked in GPFSFSAL_UP_Thread() to decide if the upcall
thread needs to exit.

Change-Id: I8cdf193ce955cd7248104741927307974d8f0958
Signed-off-by: Madhu Thorat <madhu.punjabi@in.ibm.com>
Reviewed-by: Malahal Naineni <malahal@us.ibm.com>