Update patch set 2
Patch Set 2:

I think FSAL_CEPH is broken with respect to exports.  It shouldn't be storing an export in a handle, since a handle can be reachable from multiple exports.   MDCACHE jumps through hoops to get this to sort-of work for background reaping.  I think FSAL_CEPH will have crashes/use-after-free in the presence of complicate exports and/or export addition/removal.  I think the export pointer in the FSAL_CEPH handle needs to be removed.

That said, nothing is killing the root entry of an export when it's gone.  This isn't necessarily a problem, as the entry will be reaped and recycled as other entries are created, but it might be worth killing an entry if it's last export is gone and it has no state.  Or maybe just let it get recycled.

Patch-set: 2
CC: Gerrit User 1004087 <1004087@d5d70762-12d0-45a1-890d-524b12d3f735>