Add clarification about removal of cih_remove_checked call in 0693097d6
In commit 0693097d61f33d9c8d2bd49b89d448c4f804dd0d a call to
cih_remove_checked was removed. This patch documents in comments why
that call is not necessary.

The call was not necessary because the loop that is unmapping entries
and exports will unmap the export root, and if the export root no
longer belongs to any export, it will be unhashed inside the loop.

The unhash outside the loop did nothing for export roots that only
belonged to one export, but any that belonged to two exports would
get incorrectly unhashed if one of those two exports was removed.

Change-Id: Ic596b516be0dcf7804b9ac5a1ef0976e282e6fb9
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>