FSAL_CEPH: endian-convert wire handles
The wire handle is just treated as opaque now, which means we'll end up
with different filehandles on BE and LE arches. Most of the deployments
in the field currently are LE, so settle on the fields in there being
represented as LE values so we don't change the FH unnecessarily.

Change-Id: I646c7fc3835d0602d22644e4967d458909fc816c
Signed-off-by: Jeff Layton <jlayton@redhat.com>