Update patch set 1
Patch Set 1:

This poses a problem.  If the export is part of the handle hash, then multiple handles will exist in MDCACHE for the same inode, and this will break locking.  If the export is *not* part of the hash, then MDCACHE will only keep one of the handles (they'll collide, and one will be discarded), and only one export will ever be used.

Can we just always use the current export?  It's available for every operation.

Patch-set: 1
CC: Gerrit User 1004087 <1004087@d5d70762-12d0-45a1-890d-524b12d3f735>