Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> This poses a problem.  If the export is part of the handle hash, then multiple handles will exist in MDCACHE for the same inode, and this will break locking.  If the export is *not* part of the hash, then MDCACHE will only keep one of the handles (they'll collide, and one will be discarded), and only one export will ever be used.
> 

I don't quite grok how this will break locking. The underlying inodes in this case are the same but the Inode object pointers we get back from libcephfs will be different on the different exports. To ganesha, these are completely different objects.

> Can we just always use the current export?  It's available for every operation.

No. An Inode pointer is only valid within the context of a particular ceph mount (which has a 1:1 correspondence with the export).

Patch-set: 1