Update patch set 1
Patch Set 1:

Yes. If you were running ganesha with FSAL_CEPH exports on a BE arch before, then this will change the filehandles. We'll definitely need to document that. Let me know the best place to do it and I'll write something up.

That said, the intersection of BE arches in ceph environments is pretty small so far, so I'm hoping very few people (if any) will be affected.

Patch-set: 1