Update patch set 2
Patch Set 2:

> We MUST document any limitations introduced by this, but the alternate is forcing the two exports to have the same credentials and other bits so they can share a Ceph client.

This is very similar to the situation we have when different ganesha servers in an active/active cluster are exporting the same data. The only real difference is that the filehandles are different when the exports have different export_ids, as is the case where the same data is exported via multiple exports.

In both cases, the MDS has to mediate the access between the ceph clients. That should work properly for about everything, but share locking. But, that's already something we can't handle properly between different active/active server heads, so I don't see it as a huge issue.

Patch-set: 2