ProxyV3: Fix reuse of idle-closed connections.
The V3 proxy has a pool of connections to NFS backends. If enough time has
passed though, the backend may have sent us a FIN/ACK to recycle the idle
connection. If we then try to reuse this fd, we can happily attempt to perform
write()s but then get a rude EOF on read() for the response.

Instead, we can use recv() with MSG_DONTWAIT | MSG_PEEK to basically ask the
kernel "Did the other side hang up?" before handing it out.

Testing: Manual verification that cat followed by sleep 300 and another cat of a
different file (to avoid caching) no longer results in an EOF on the "read the
header response". Instead, we get a proper RST on close(fd) and then re-opening
with a FullDebug log message like:

 proxyv3_getfdentry :FSAL :DEBUG :Socket 7 was closed by the backend.

Change-Id: I0bdf8751fa8dabc599696b5b49b9ea2e4dda8676
Signed-off-by: Solomon Boulos <boulos@google.com>