Fix PNFS flags in EXCHANGE_ID
NFSv4.1 RFC provides the list of combinations that are acceptable in case
of PNFS. However, all these restrictions are only applied to server, the
client is allowed to send any combination when requesting.
This change fixes the EXCHANGE_ID code to make sure that the server always
returns acceptable PNFS flags even if client has request invalid
combination.

Change-Id: Iba05fb9729beda8bc30c957f3a3f153a37c8554c
Signed-off-by: Sriram Patil <sriramp@vmware.com>