move run dir to /run
/run has been preferred over /var/run for a while.
The main motivation for this change is that /run is not dependent on
SYSSTATEDIR, and under some distributions (e.g. nixos) led to
improbable run dir.

Also, for systemd distributions, /run/ganesha has been created by
systemd for a while (via the RuntimeDirectory= setting), so embrace
the dynamic directory creation for all service files

Change-Id: I83b3871d5dd1984534487e6e237dbe8ae0eefe00
Signed-off-by: Dominique Martinet <dominique.martinet@cea.fr>