Update patch set 4
Patch Set 4:

> Ganesha creates the PID file itself, and it's hard coded to create it in GANESHA_PIDFILE_PATH.  Currently, this is inside the PREFIX set by CMake.  With this change, it will always be in /run, and so outside the prefix.

Well, that it is inside the prefix precisely was my problem on nixos ;) As systemd names it, it really is a "runtime" directory that has nothing to do with the install prefix.
At the very least even if I restore some option for both paths I would want the default to be /run outside of the prefix.

I had seen the file specified in some service file (-p option) but you are right it appears to only be done in the gpfs one, it's not general...
otoh none of these service files are configurable either so I think making it a compile time option would be somewhat self-satisfactory as it would be difficult to use... I think it's easier to both pass the option in the service along with where the service should expect it to be than change the compile-time default.
Anyway it doesn't cost much to make it a compile time option, will refresh the patch with a RUNTIMEDIR in cmake defaulting to /run.

Patch-set: 4