Update patch set 1
Patch Set 1:

A better fix for this might be to make sure ERR_FSAL_STALE is handled properly to kill the entry. I'm not sure if this is the first detection point in all cases or if we should have another.

Patch-set: 1
Reviewer: Gerrit User 1003571 <1003571@d5d70762-12d0-45a1-890d-524b12d3f735>