Simplify fsal_ds_handle
This handle just has the lifetime of a single COMPOUND and other than
the remote possibility it could be subject to SAVEFH/RESTOREFH can
be very simply handled so it doesn't need a refcount or to be on a list
of handles for the fsal_pnfs_ds.

Move the dsh_ops into the fsal_pnfs_ds. There really is no need for
handle specific ops. This makes the default struct fsal_ds_handle
empty, however, FSALs will define their own private structure that
extends it.

Change-Id: I0c83f1f8e657cec533da26dd92f624a11e2ca35e
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>