Fix open file limit detection on macOS
The hard limit reported by getrlimit on macOS is not correct. It has to
be brought down to MAX_OPEN if it is higher than it.

Create a new sys_resource module which differs by OS. I did not add this
to the subr module since that functionality is only needed for POSIX
filesystem support (e.g. VFS FSAL). On the other hand, the getrlimit for
open files is needed by all FSALs because it is required by mdcache.

Change-Id: Icd685ef907239fe3e1f32168b3a5ee408798a371
Signed-off-by: Matthew DeVore <matvore@google.com>