Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-2
> 
> NAK - See discussion here: https://lists.nfs-ganesha.org/archives/list/devel@lists.nfs-ganesha.org/thread/UIJLKD66O4BYKFS23UFKYVPUO33XPN5B/
> 
> The translation of ENXIO at end of file is to match the NFSv4.2 protocol which doesn't exactly match the POSIX defined behavior for SEEK_HOLE past end of file.
> 
> We made this coding choice explicitly to enable expected behavior on the client and to be compatible with knfsd.
> 
> Unless something has changed in the protocol spec or how the Linux client and/or knfsd handle this situation, this change can not be accepted.

I looked this up before acking, and I think this is correct.  7862 says:

If the sa_offset is beyond the end of the file, then SEEK
   MUST return NFS4ERR_NXIO.

And man lseek says:

ENXIO  whence is SEEK_DATA or SEEK_HOLE, and the file offset is beyond the end of the file.

I think this means that ENXIO from lseek should be returned to the client.

Patch-set: 1
Reviewer: Gerrit User 1004087 <1004087@d5d70762-12d0-45a1-890d-524b12d3f735>