Update patch set 1
Patch Set 1:

I wonder if something has changed since I first wrote and tested the code 2 years ago. I'll have to research more. It may also be a lack of clarity of what "beyond the end of file" means. If I recall, the issue was seeking a hole at an offset exactly at the end of file. The client and testing and documentation seemed to expect to find a hole after the final data in a file was correct.

I don't want to change code that was explicitly coded that way after a discussion when we were doing the implementation without further discussion.

Patch-set: 1
Reviewer: Gerrit User 1003571 <1003571@d5d70762-12d0-45a1-890d-524b12d3f735>