Update patch set 3
Patch Set 3: Code-Review-1

So, this test in nfs_sparse is wrong:

https://github.com/thombashi/NFStest/blob/be1c8634a63a749baf7b8c377dcd92a551af3ea4/test/nfstest_sparse#L418

In that case, which is a SEEK_DATA within a hole at the end of the file, should *not* return ENXIO according to the RFC.  The RFC says:

If the server cannot find a corresponding sa_what, then
   the status will still be NFS4_OK, but sr_eof would be TRUE.

So, we should be returning NFS4_OK with EOF, which is what the code appears to do.   I think Ganesha is correct, modulo a race in checking the file length on a file being shrunk.

Patch-set: 3
Reviewer: Gerrit User 1004087 <1004087@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1