PROXY_V3: work around bindresvport threadsafety.
To get a privileged port, the V3 proxy's rpc code
asks bindresvport_sa to hand us one. Modern versions
of glibc (since 2.17) are thread-safe, but many existing
implementations (including libntirpc) just use a local
static variable for the "choose a port". The V4 proxy
calls the related rresvport_af but is holding a lock,
so it's at least single threaded amongst itself.

The race here was causing the client port to sometimes be an
unprivileged port, so knfsd was correctly returning a
permission error.

While debugging this with very few max connections, I
realized that the exponential backoff code wasn't capping
the delay. With 100 max iterations, the delay would easily
jump into the several seconds (and as there is still no
jitter, all the service threads hit the delay together). To
make it a little better, this change also caps that at 10 ms.

Change-Id: I5a9f60451cf9154130a6f3fc21979cbe6e6d2049
Fixes: #580
Signed-off-by: Solomon Boulos <boulos@google.com>