Set root of FSAL_MEM exports to writeable by all
This allows clients which set the owner of the root directory of the
export to one user (including super user) to still allow writes by other
users, and potentially allow multiple users to own files on the share.

This is "insecure" but more flexible, which is fine for FSAL_MEM exports
since FSAL_MEM is not intended for production use.

Change-Id: Ifb5ac8488cc705bf0ca453ceef05a6c19996ded6
Signed-off-by: Matthew DeVore <matvore@google.com>