ProxyV3: Issue last-chance GETATTR in READDIRPLUS.
NetApp's .snapshot directories apparently return no attributes
for a READDIRPLUS request. It's easy enough to issue a last-chance
"try with GETATTR" to fill in the attributes.

I tested this by just always doing the last-chance path and verified
it on both cthon04, manual testing (of ls), and the repeated rsync
from my earlier patch. I have *not* tested this against a NetApp
filer, but expect leffler@google.com to do so.

Change-Id: Ia25a5d6b2b8966bad4f173c3d62f2e87a8b68adb
Signed-off-by: Solomon Boulos <boulos@google.com>