Update patch set 6
Patch Set 6:

Hmm, this isn't getting folks excited.

I'm looking at the possibility for an alternate structuring:

FSAL {
    Name = VFS;
    EXPORT {
        ...
    }
    EXPORT {
        ...
    }
}

The problem with that is that the EXPORT sub-blocks would get committed before the FSAL block and it may be tricky for the EXPORT to get the FSAL parse node (well, we would just have to add a config helper to fetch the parent from the void *node). In theory, this would allow a group of EXPORT blocks to share common FSAL block parameters without totally hacking up the existing code.

The EXPORT commit would just have to detect that EXPORT was a sub-block instead of a top level block, and proceed accordingly. Actually, it might just have to detect that there is no fsal_export...

This would NOT allow the dbus add export command to add such an export but SIGHUP export reload would still work.

Multiple FSAL blocks with EXPORT sub-blocks would be allowed.

Patch-set: 6
Reviewer: Gerrit User 1003571 <1003571@d5d70762-12d0-45a1-890d-524b12d3f735>