Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > Another option would be to change the enable_UDP option from boolean to an enum, allowing some synonyms, so the values might be:
> > > 
> > > true or yes
> > > false or no
> > > mnt_only
> > > 
> > > That would make the config more clear.
> > 
> > Thanks Frank, will update the patch accordingly.
> > one point: If anyone is already using this option in config, they will need to update it in the upgrade flow. Is this acceptable?
> 
> As long as it accepts the current synonyms for True and False, they shouldn't need to change config, I think.

Yes, that was my intent, make the underlying variable tri-state but keep the existing values for on and off (check all the synonyms for true and false that are used) and add a third mid-state. Then existing config files still set the value expected while new configs use the new value.

Patch-set: 1