mem_handle: Do not grab lock on destroyed handle
mem_clean_all_dirents is only called in one place, and it is after the
object handle has been destroyed. Hence, there is no valid lock to
acquire. This cleanup also happens after the final reference count to
the object is released, so there should be no danger of concurrent
modification.

This fixes a crash when running st_rename.testDirToDir in pynfs on
macOS.

Change-Id: I18b9cabcb31d82463279cb6232e0b26bbea16d4d
Signed-off-by: Matthew DeVore <matvore@google.com>