Update patch set 1
Patch Set 1:

The problem is that the sentinel ref *is* the hash ref, so it needs to be removed when it's removed from the hash.

mdc_up_try_release() should be fine; it explicitly takes an extra ref before trying to put the sentinel ref, so it also needs to release that one at the end.  It also doesn't look like up-calls were involved in the backtrace you posted.

Patch-set: 1
CC: Gerrit User 1004087 <1004087@d5d70762-12d0-45a1-890d-524b12d3f735>