Update patch set 3
Patch Set 3:

Oh, and one more note: when pushing new versions of patches to Gerrithub that were pushed as a "branch" please push the whole branch again. Sometimes I pull the branch or like today, I open the last patch first, and then navigated to the first patch, which then was the original version of this patch... Luckily I caught it. Sometimes I don't catch that I'm downloading not the latest version of the patch...

Pushing branches is nice because if there are more than 2 or 3 patches, it saves a lot of time to pull the whole branch rather than cherry pick each patch.

Patch-set: 3
Reviewer: Gerrit User 1003571 <1003571@d5d70762-12d0-45a1-890d-524b12d3f735>