ProxyV3: Add an optional mount_path parameter.
I had been using the proxy just to mirror a backend server. Bjorn Leffler
pointed out that it would be handy to allow the proxy to also have
the ability to export a different path than the mount path, particularly if
two backend servers use the same export:

EXPORT {
  Export_Id = 1;
  Path = "/server1";
  Pseudo = "/server1";
  FSAL {
     Name = "PROXY_V3";
     Srv_addr = "192.168.1.2";
     Mount_Path = "/export";
  }
}

EXPORT {
  Export_Id = 2;
  Path = "/server2";
  Pseudo = "/server2";
  FSAL {
     Name = "PROXY_V3";
     Srv_addr = "192.168.1.3";
     Mount_Path = "/export";
  }
}

I left it to defaulting to "mirroring" mode (use the export path), but this
seemed like a nice feature request.

Change-Id: I9468733f13dcf37b3eb856a96a02061e80ccf4a0
Signed-off-by: Solomon Boulos <boulos@google.com>