Update patch set 1
Patch Set 1:

Hmm, what is the consequence of not fetching the permission set? It seems like if there is a real chance of failure there we should do something. If we really don't need to do anything, we should have  a comment to that effect.

Patch-set: 1
Reviewer: Gerrit User 1003571 <1003571@d5d70762-12d0-45a1-890d-524b12d3f735>