CEPH: Introduce host_to_key and fix up use of host handles and handle-key
Fixes GitHub Issue 661

After clarifying how the various handle forms are used, and noting that
CEPH was missing host_to_key, it's clear that wire_to_host MUST produce
a host handle (without the export_id), not a handle-key. If the handle-key
(with the export_id) is needed, MDCACHE will call host_to_key.

The create_handle method is passed a host handle NOT a handle-key so it
must NOT expect the export_id to be there.

Change-Id: I4e06731d4c27da5a89eda7c08b79f93de84a4855
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>