Allow host_to_key to return a key that is larger than the host handle
Some FSALs may need to extend the host handle to create the handle-key.
For example FSAL_CEPH has a need to have each fsal_obj_handle belong to
a specific export even when the object might appear in multiple exports.
To accomplish this, FSAL_CEPH adds the export_id to the host handle to
produce the handle-key.

There is a subsequent patch to the FSAL_CEPH code to use this new
contract and resolve the incorrect way it added the export_id that
didn't add it everywhere.

The host_to_key method may add AT MOST FSAL_KEY_EXTRA_BYTES. Callers
of host_to_key MUST allocate this extra space.

Change-Id: I70850dd7c14697549b8fc2c7545a07844ac67ac0
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>