Update patch set 1
Patch Set 1:

Hmm, that makes sense, but is there an issue with squashing? The original creds won't be squashed which could cause different behavior, on the other hand, considering that Ganesha expects the underlying file system to do permission checking for directory operations (and some file operations), the squash settings between Ganesha and the back end server SHOULD be the same.

Patch-set: 1
CC: Gerrit User 1003571 <1003571@d5d70762-12d0-45a1-890d-524b12d3f735>