Update patch set 1
Patch Set 1:

Right, this is actually "If MDCACHE could delegate (with caching?) to the fsal below, the access check would happen on the backend".

I considered doing some sort of "oh, maybe I should trim the list" but since the *original* request is there and I'm a *proxy*, I should actually pretend that the request reached the backend unmodified.

WDYT?

Patch-set: 1