Clean up log config
Make Default_Log_Level useful. Now if components are not specified, they will
be set to the Default_Log_Level. The Default_Log_Level is preset from the -N
option.

If component ALL is specified along with other components, any component not
explicitly mentioned will be set to the level ALL was specified as. If ALL is
removed, Default_Log_Level takes over, thus ALL can be unset also.

If a component was explicitly set, and then removed, the component will be
reset to Default_Log_Level.

Log level changes are always logged.

RPC_Debug_Level is cleaned up so it is set as now documented.

The logging documentation is improved.

Change-Id: I39648fa626b6307d47151ac171806e2b1aeb0edc
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>