NFS: convert LISTXATTRS to RFC 8276
Convert the LISTXATTRS code to convert with RFC 8276. Note that we also
change the listxattrs API a bit. The old code had nfs4_op_listxattr
allocate a (largeish) array based on an estimate using the maxcount
passed in by the client. That covered most reasonable cases, but might
not have worked well at the edges, and it was a bit odd in that the FSAL
still had to allocate the char arrays for utf8string_val.

The new API passes in a pointer to a zeroed out xattrlist4 that the FSAL
must fill out itself. The FSAL can then allocate an array that is
correctly sized, based on the xattrs it finds and the args from the
client.

Change-Id: I6b742dc04f6ae9fc434bd1bdcdaff04e8b8109ea
Signed-off-by: Jeff Layton <jlayton@redhat.com>